+++ Infin8.Coapp.Dto/MemberDashBoardTD.cs
@@ -2,8 +2,8 @@
 {
     public  class MemberDashBoardTD
     {
-        public int Mem_Id { get; set; }
-        public int TD_Id { get; set; }
+        public decimal Mem_Id { get; set; }
+        public decimal TD_Id { get; set; }
         public int TDScheme_Id { get; set; }
         public string? TDScheme_Name { get; set; }
         public string? TD_No { get; set; }
